Change: AppDelegate in Extension auslagern
